### 2023-01-09 e5b1694
Adding JOSS folder (#319)
* bumping SEMVER to v3.66

* rebase main

* WIP: fixing posit/literal logic operators for specialized fast posits

* adding JOSS folder for publication

* bug fix on max shift calculation in LNS

* adding first and second base non-type params to mdlns as first test

* bug fix on mdlns

* adding JOSS pdf workflow and paper edits

* fixing input and output paths of JOSS paper
