+++ joss/paper.md
@@ -1,0 +1,159 @@
+---
+title: 'Universal Numbers Library: Multi-format Variable Precision Arithmetic Library'
+tags:
+  - C++
+  - numerical analysis
+  - deep learning
+  - machine learning
+  - floating-point representations
+  - mixed-precision
+authors:
+  - name: E. Theodore L. Omtzigt
+    orcid: 0000-0003-0194-951X
+    equal-contrib: true
+    affiliation: 1
+  - name: James Quinlan
+    orcid: 0000-0002-2628-1651
+    equal-contrib: true 
+    corresponding: true
+    affiliation: 2
+affiliations:
+ - name: Stillwater Supercomputing, Inc, USA
+   index: 1
+ - name: School of Mathematical and Physical Sciences, University of New England, USA
+   index: 2
+date: 15 December 2022
+bibliography: references.bib
+---
+
+# Summary
+
+*Universal Numbers Library*, or *Universal* for short, is a self-contained C++ header-only template library that contains implementations of many number representations and standard arithmetic on arbitrary configuration integer and real numbers [@omtzigt:2020].  In particular, the library includes integers, decimals, fixed-points, rationals, linear floats, tapered floats, logarithmic, SORNs, interval, level-index, and adaptive-precision binary and decimal integers and floats, each offering a verification suite.  
+
+The primary pattern using a posit number type as example, is:
+
+```cpp
+#include <universal/number/posit/posit.hpp>
+
+template<typename Real>
+Real MyKernel(const Real& a, const Real& b) {
+    return a * b;  // replace this with your kernel computation
+}
+
+constexpr double pi = 3.14159265358979323846;
+
+int main() {
+    using Real = sw::universal::posit<32,2>;  
+
+    Real a = sqrt(2);
+    Real b = pi;
+    std::cout << "Result: " << MyKernel(a, b) << std::endl;
+}
+```
+
+
+*Universal* delivers software and hardware co-design capabilities to develop low and mixed-precision algorithms for reducing energy consumption in signal processing, Industry 4.0, machine learning, robotics, and high-performance computing applications [@omtzigt:2022].  The package includes command-line tools for visualizing and interrogating numeric encodings, an interface for setting and querying bits, and educational examples showcasing performance gain and numerical accuracy with the different number systems.  In addition, a Docker container is available to experiment without cloning and building from the source code.
+
+```
+$ docker pull stillwater/universal
+$ docker run -it --rm stillwater/universal bash
+```
+
+*Universal* started in 2017 as a bit-level arithmetic reference implementation of the evolving unum Type III (posit and valid) standard.  However, the demands for supporting various number systems, such as adaptive-precision integers to solve large factorials, adaptive-precision floats to act as Oracles, or comparing linear and tapered floats provided the opportunity to create a complete platform for numerical analysis and computational mathematics.  As a result, several projects have leveraged *Universal*, including Matrix Template Library (MTL4), Geometry + Simulation Modules (G+SMO), Bembel, a fast IGA BEM solver, and the Odeint ODE solver.
+
+The default build configuration will produce the command line tools, a playground, and educational and application examples.  It is also possible to construct the full regression suite across all the number systems.  For instance, the shortened output for the commands `single` and `single 1.23456789` are below.
+
+```bash
+$ single
+min exponent                                           -125
+max exponent                                            128
+radix                                                     2
+radix digits                                             24
+min                                             1.17549e-38
+max                                             3.40282e+38
+lowest                                         -3.40282e+38
+epsilon (1+1ULP-1)                              1.19209e-07
+round_error                                             0.5
+denorm_min                                       1.4013e-45
[…]
