+++ include/universal/number/mdlns/mdlns_impl.hpp
@@ -17,8 +17,8 @@
 namespace sw { namespace universal {
 		
 // convert a floating-point value to a specific mdlns configuration. Semantically, p = v, return reference to p
-template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
-inline mdlns<nbits, rbits, bt, xtra...>& convert(const triple<nbits, bt>& v, mdlns<nbits, rbits, bt, xtra...>& p) {
+template<unsigned nbits, unsigned rbits, unsigned firstBase, unsigned secondBase, typename bt, auto... xtra>
+inline mdlns<nbits, rbits, firstBase, secondBase, bt, xtra...>& convert(const triple<nbits, bt>& v, mdlns<nbits, rbits, firstBase, secondBase, bt, xtra...>& p) {
 	if (v.iszero()) {
 		p.setzero();
 		return p;
@@ -30,25 +30,25 @@ inline mdlns<nbits, rbits, bt, xtra...>& convert(const triple<nbits, bt>& v, mdl
 	return p;
 }
 
-template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
-inline mdlns<nbits, rbits, bt, xtra...>& minpos(mdlns<nbits, rbits, bt, xtra...>& lminpos) {
+template<unsigned nbits, unsigned rbits, unsigned firstBase, unsigned secondBase, typename bt, auto... xtra>
+inline mdlns<nbits, rbits, firstBase, secondBase, bt, xtra...>& minpos(mdlns<nbits, rbits, firstBase, secondBase, bt, xtra...>& lminpos) {
 	return lminpos;
 }
-template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
-mdlns<nbits, rbits, bt, xtra...>& maxpos(mdlns<nbits, rbits, bt, xtra...>& lmaxpos) {
+template<unsigned nbits, unsigned rbits, unsigned firstBase, unsigned secondBase, typename bt, auto... xtra>
+mdlns<nbits, rbits, firstBase, secondBase, bt, xtra...>& maxpos(mdlns<nbits, rbits, firstBase, secondBase, bt, xtra...>& lmaxpos) {
 	return lmaxpos;
 }
-template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
-mdlns<nbits, rbits, bt, xtra...>& minneg(mdlns<nbits, rbits, bt, xtra...>& lminneg) {
+template<unsigned nbits, unsigned rbits, unsigned firstBase, unsigned secondBase, typename bt, auto... xtra>
+mdlns<nbits, rbits, firstBase, secondBase, bt, xtra...>& minneg(mdlns<nbits, rbits, firstBase, secondBase, bt, xtra...>& lminneg) {
 	return lminneg;
 }
-template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
-mdlns<nbits, rbits, bt, xtra...>& maxneg(mdlns<nbits, rbits, bt, xtra...>& lmaxneg) {
+template<unsigned nbits, unsigned rbits, unsigned firstBase, unsigned secondBase, typename bt, auto... xtra>
+mdlns<nbits, rbits, firstBase, secondBase, bt, xtra...>& maxneg(mdlns<nbits, rbits, firstBase, secondBase, bt, xtra...>& lmaxneg) {
 	return lmaxneg;
 }
 
 // template class representing a value in scientific notation, using a template size for the number of fraction bits
-template<unsigned _nbits, unsigned _rbits, typename bt = uint8_t, auto... xtra>
+template<unsigned _nbits, unsigned _rbits, unsigned firstBase = 2, unsigned secondBase = 3, typename bt = uint8_t, auto... xtra>
[…]
