### 2023-01-10 #5058 @Ravenwater
@danielskatz weird, it is in the joss subdirectory in the main branch:  .../joss/paper.md

![image](https://user-images.githubusercontent.com/466641/211553584-e09a9aae-2a73-4877-a642-ee1050e548fe.png)

We had some issues in the submission process: it did not like a free string "computer arithmetic" in the paper subject and I didn't realize that the UI has a popup that provides the valid selections. As I was trying to figure this out, I copy and pasted the descriptor, and when I pasted it back into the final submission that has the right subject attribute, I missed the "No", in "No conflicts of interest..."   I presumed that looked very odd for a submission, hence my note here.
