### 2023-01-10 e21d5db
adding JOSS submission to the README
