### 2023-01-10 #5058 @danielskatz
ok, I'm not sure what happened.  Please go ahead and add the DOIs to the references that are missing them - see the bib example of the sample paper in the JOSS docs if you need an example.
