### 2023-01-10 #5058 @danielskatz
@Ravenwater - it looks like that joss directory was just added 13 hours ago, which was after the initial submission and the initial failures of the JOSS system (14 hours ago).
