### 2023-01-10 #5058 @jamesquinlan
@danielskatz FYI, waiting for @Ravenwater to push the new BibTeX file with doi to the main repo.  As for reviewers, jordigh and wrathmatics from top of the list seem like a match.  @Ravenwater, any suggested reviewers?
