### 2023-01-10 426b12f
updating JOSS submission (#320)
* bumping SEMVER to v3.67

* moving CI to new dev branch

* disabling the JOSS darft paper workflow

* adding JOSS submission to the README

* updating bibtex with DOIs
