+++ docker/build_release_container.sh
@@ -5,7 +5,7 @@
 # example would be to strace an executable to find its dependencies
 
 MAJOR=v3
-MINOR=66
+MINOR=67
 VERSION="$MAJOR.$MINOR"
 
 if [[ $# == 0 ]]; then
