+++ joss/references.bib
@@ -2,13 +2,14 @@
 %%% BibTeX-file{
 %%% 	title		= "Universal Numbers Library"
 %%%     author          = "Omtzigt, E.T.L & Quinlan, J.",
-%%%     version         = "1.00",
-%%%     date            = "09 Jan 2023",
+%%%     version         = "1.01",
+%%%     date            = "10 Jan 2023",
 %%%     filename        = "references.bib",
 %%%     keywords        = "Universal Numbers Library"
 %%%     AMS             = "65G50, 65F10",
 %%%     }
 %%% =========================================================
+
 
 
 @inproceedings{carmichael:2019,
@@ -17,6 +18,7 @@
   booktitle={2019 Design, Automation \& Test in Europe Conference \& Exhibition (DATE)},
   pages={1421--1426},
   year={2019},
+  doi = {10.23919/date.2019.8715262},
   organization={IEEE}
 }
 
@@ -29,6 +31,7 @@
   number={2},
   pages={A817--A847},
   year={2018},
+  doi = {10.1137/17m1140819},
   publisher={SIAM}
 }
 
@@ -41,6 +44,7 @@
   number={2},
   pages={29--35},
   year={2021},
+  doi={10.1109/mm.2021.3061394},
   publisher={IEEE}
 }
 
@@ -60,6 +64,7 @@
   volume={13253},
   pages={14},
   year={2022},
+  doi={10.1007/978-3-031-09779-9_2},
   organization={Springer Nature}
 }
 
@@ -70,6 +75,7 @@
   booktitle={SC18: International Conference for High Performance Computing, Networking, Storage and Analysis},
   pages={603--613},
   year={2018},
+  doi={10.1109/sc.2018.00050},
   organization={IEEE}
 }
 
@@ -79,6 +85,7 @@
   author={Haidar, Azzam and Wu, Panruo and Tomov, Stanimire and Dongarra, Jack},
   booktitle={Proceedings of the 8th Workshop on Latest Advances in Scalable Algorithms for Large-Scale Systems},
   pages={1--8},
+  doi={10.1145/3148226.3148237},
   year={2017}
 }
 
@@ -89,6 +96,7 @@
   booktitle={International Conference on Computational Science},
   pages={586--600},
   year={2018},
+  doi={10.1007/978-3-319-93698-7_45},
   organization={Springer}
 }
 
@@ -101,6 +109,7 @@
   number={4},
   pages={A2536--A2551},
   year={2019},
+  doi={10.1137/18m1229511},
   publisher={SIAM}
 }
 
@@ -114,8 +123,9 @@
   pages={},
   year={2018},
   publisher={Intel}, 
-  url={https://www.intel.com/content/dam/develop/external/us/en/documents/bf16-hardware-numerics-definition-white-paper.pdf}
+  url={https://tinyurl.com/y8ybct4}
 }
+% https://www.intel.com/content/dam/develop/external/us/en/documents/bf16-hardware-numerics-definition-white-paper.pdf
 
 
 @article{kharya:2020,
@@ -133,6 +143,7 @@
   booktitle={Conference on Next Generation Arithmetic},
   pages={100--116},
   year={2022},
+  doi={10.1007/978-3-031-09779-9_7},
   organization={Springer}
 }
 
