+++ .github/workflows/cmake.yml
@@ -2,7 +2,7 @@ name: CMake
 
 on:
   push:
-    branches: [ v3.66, main ]
+    branches: [ v3.67, main ]
   pull_request:
     branches: [ main ]
 
