### 2023-01-10 #5058 @danielskatz
👋 @wrathematics - Would you be able to review this JOSS submission?
