### 2023-01-10 #5058 @Ravenwater
updated bibtex with James DOI edits:
