### 2023-01-10 #5058 @danielskatz
> editorialbot commands need to be the first entry in a new comment.
