### 2023-01-10 #5058 @Ravenwater
@danielskatz the JOSS process is awesome!!! tooling, collaboration, and automation are all well done and are a joy to work with.
