### 2023-01-10 #5058 @Ravenwater
@danielskatz I know what went wrong the first time around: we have a Github actions CI system that takes about an hour to run the regression suite. I submitted the JOSS via a branch waiting for the CI to complete and be green before merging to main. I submitted the JOSS paper before the CI was done and the branch was merged. My bad.
