### 2023-01-11 bdfe302
changing mdlns to lns2b to simplify template API
