+++ include/universal/number/mdlns/mdlns_impl.hpp
@@ -1,5 +1,5 @@
 #pragma once
-// mdlns_impl.hpp: implementation of an arbitrary logarithmic number system configuration
+// lns2b_impl.hpp: implementation of a fixed-size, arbitrary configuration 2-base logarithmic number system configuration
 //
 // Copyright (C) 2017-2022 Stillwater Supercomputing, Inc.
 //
@@ -12,13 +12,13 @@
 #include <universal/internal/abstract/triple.hpp>
 #include <universal/number/shared/specific_value_encoding.hpp>
 #include <universal/behavior/arithmetic.hpp>
-#include <universal/number/mdlns/mdlns_fwd.hpp>
+#include <universal/number/lns2b/lns2b_fwd.hpp>
 
 namespace sw { namespace universal {
 		
-// convert a floating-point value to a specific mdlns configuration. Semantically, p = v, return reference to p
-template<unsigned nbits, unsigned rbits, unsigned firstBase, unsigned secondBase, typename bt, auto... xtra>
-inline mdlns<nbits, rbits, firstBase, secondBase, bt, xtra...>& convert(const triple<nbits, bt>& v, mdlns<nbits, rbits, firstBase, secondBase, bt, xtra...>& p) {
+// convert a floating-point value to a specific lns2b configuration. Semantically, p = v, return reference to p
+template<unsigned nbits, unsigned fbbits, typename bt, auto... xtra>
+inline lns2b<nbits, fbbits, bt, xtra...>& convert(const triple<nbits, bt>& v, lns2b<nbits, fbbits, bt, xtra...>& p) {
 	if (v.iszero()) {
 		p.setzero();
 		return p;
@@ -30,26 +30,26 @@ inline mdlns<nbits, rbits, firstBase, secondBase, bt, xtra...>& convert(const tr
 	return p;
 }
 
-template<unsigned nbits, unsigned rbits, unsigned firstBase, unsigned secondBase, typename bt, auto... xtra>
-inline mdlns<nbits, rbits, firstBase, secondBase, bt, xtra...>& minpos(mdlns<nbits, rbits, firstBase, secondBase, bt, xtra...>& lminpos) {
+template<unsigned nbits, unsigned fbbits, typename bt, auto... xtra>
+inline lns2b<nbits, fbbits, bt, xtra...>& minpos(lns2b<nbits, fbbits, bt, xtra...>& lminpos) {
 	return lminpos;
 }
-template<unsigned nbits, unsigned rbits, unsigned firstBase, unsigned secondBase, typename bt, auto... xtra>
-mdlns<nbits, rbits, firstBase, secondBase, bt, xtra...>& maxpos(mdlns<nbits, rbits, firstBase, secondBase, bt, xtra...>& lmaxpos) {
+template<unsigned nbits, unsigned fbbits, typename bt, auto... xtra>
+lns2b<nbits, fbbits, bt, xtra...>& maxpos(lns2b<nbits, fbbits, bt, xtra...>& lmaxpos) {
 	return lmaxpos;
 }
-template<unsigned nbits, unsigned rbits, unsigned firstBase, unsigned secondBase, typename bt, auto... xtra>
-mdlns<nbits, rbits, firstBase, secondBase, bt, xtra...>& minneg(mdlns<nbits, rbits, firstBase, secondBase, bt, xtra...>& lminneg) {
+template<unsigned nbits, unsigned fbbits, typename bt, auto... xtra>
+lns2b<nbits, fbbits, bt, xtra...>& minneg(lns2b<nbits, fbbits, bt, xtra...>& lminneg) {
 	return lminneg;
 }
[…]
