+++ include/universal/number/lns2b/lns2b_traits.hpp
@@ -1,7 +1,7 @@
 #pragma once
-//  mdlns_traits.hpp : traits for logarithmic number system
+//  lns2b_traits.hpp : traits for 2-base logarithmic number system
 //
-// Copyright (C) 2017-2022 Stillwater Supercomputing, Inc.
+// Copyright (C) 2022-2023 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #include <universal/traits/integral_constant.hpp>
@@ -10,21 +10,21 @@ namespace sw { namespace universal {
 
 // define a trait for lns types
 template<typename _Ty>
-struct is_mdlns_trait
+struct is_lns2b_trait
 	: false_type
 {
 };
 
-template<unsigned nbits, unsigned rbits, unsigned firstBase, unsigned secondBase, typename BlockType, auto... xtra>
-struct is_mdlns_trait< mdlns<nbits, rbits, firstBase, secondBase, BlockType, xtra...> >
+template<unsigned nbits, unsigned fbbits, typename BlockType, auto... xtra>
+struct is_lns2b_trait< lns2b<nbits, fbbits, BlockType, xtra...> >
 	: true_type
 {
 };
 
 template<typename _Ty>
-constexpr bool is_mdlns = is_mdlns_trait<_Ty>::value;
+constexpr bool is_lns2b = is_lns2b_trait<_Ty>::value;
 
 template<typename _Ty>
-using enable_if_mdlns = std::enable_if_t<is_mdlns<_Ty>, _Ty>;
+using enable_if_lns2b = std::enable_if_t<is_lns2b<_Ty>, _Ty>;
 
 }} // namespace sw::universal
