+++ static/mdlns/README.md
@@ -1,5 +1,0 @@
-# MDLNS: Multi-Dimensional Logarithmic Number System
-
-A number X, is represented by a composition of logarithms, a, b, c, ... , of its absolute value:
-
-X -> {s, 2^a * 3^b * 5^c ... )

[…]
