### 2023-01-11 21cb6a1
WIP: stitching back together the nw lns2b type
