+++ static/lns2b/api/api.cpp
@@ -4,7 +4,8 @@
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #include <universal/utility/directives.hpp>
-#include <universal/number/mdlns/mdlns.hpp>
+#include <universal/number/lns2b/lns2b.hpp>
+#include <universal/number/lns2b/table.hpp>
 #include <universal/number/cfloat/cfloat.hpp>  // bit field comparisons
 #include <universal/verification/test_suite.hpp>
 
@@ -40,24 +41,28 @@ int main()
 try {
 	using namespace sw::universal;
 
-	std::string test_suite = "mdlns API demonstration";
+	std::string test_suite = "lns2b API demonstration";
 	std::string test_tag = "api";
 	bool reportTestCases = false;
 	int nrOfFailedTestCases = 0;
 
 	ReportTestSuiteHeader(test_suite, reportTestCases);
 
-	float x{ 1.0f };
-	bases(x, 2, 3, 5, 7, 9);
-	return 0;
+	// float x{ 1.0f };
+	// bases(x, 2, 3, 5, 7, 9);
+
+	// generate a value table for lns2b<5,2>
+	GenerateLns2bTable<5,2>(std::cout);
+//	lns2b<5, 2> l(1);
+//	std::cout << to_binary(l) << " : " << l << " : " << color_print(l) << '\n';
 
 	// important behavioral traits
-//	ReportTrivialityOfType<mdlns<8, 2>>();  // TODO: type_tag fails for mdlns
+//	ReportTrivialityOfType<lns2b<8, 2>>();  // TODO: type_tag fails for lns2b
 
 	// default behavior
 	{
-		std::cout << "+---------    default mdmdlns bahavior   --------+\n";
-		using Real = mdlns<8, 3>;
+		std::cout << "+---------    default lns2b bahavior   --------+\n";
+		using Real = lns2b<8, 3>;
 		Real a(1.0f), b(1.0f), c;
 //		ArithmeticOperators<Real>(a, b);
 		a = 1;  // integer assignment
@@ -69,30 +74,30 @@ try {
 	// configuration
 	{
 		std::cout << "+---------    arithmetic operators with explicit alignment bahavior   --------+\n";
-//		using mdlns16 = mdlns<16, 5, std::uint16_t>;
-//		ArithmeticOperators<mdlns16>(1.0f, 1.0f);
+//		using lns2b16 = lns2b<16, 5, std::uint16_t>;
+//		ArithmeticOperators<lns2b16>(1.0f, 1.0f);
 
-//		using mdlns24 = mdlns<24, 5, std::uint32_t>;
-//		ArithmeticOperators<mdlns24>(1.0f, 1.0f);
+//		using lns2b24 = lns2b<24, 5, std::uint32_t>;
+//		ArithmeticOperators<lns2b24>(1.0f, 1.0f);
 	}
 
 	{
-		std::cout << "+---------    Dynamic ranges of mdlns<> configurations   --------+\n";
-//		std::cout << dynamic_range(mdlns< 4, 2>()) << '\n';
-//		std::cout << dynamic_range(mdlns< 8, 3>()) << '\n';
-//		std::cout << dynamic_range(mdlns<12, 4>()) << '\n';
-//		std::cout << dynamic_range(mdlns<16, 5>()) << '\n';
-//		std::cout << dynamic_range(mdlns<20, 6>()) << '\n';
+		std::cout << "+---------    Dynamic ranges of lns2b<> configurations   --------+\n";
+//		std::cout << dynamic_range(lns2b< 4, 2>()) << '\n';
[…]
