+++ include/universal/number/lns2b/lns2b_impl.hpp
@@ -48,20 +48,20 @@ lns2b<nbits, fbbits, bt, xtra...>& maxneg(lns2b<nbits, fbbits, bt, xtra...>& lma
 }
 
 // template class representing a value in scientific notation, using a template size for the number of fraction bits
-template<unsigned _nbits, unsigned _rbits, unsigned firstBase = 2, unsigned secondBase = 3, typename bt = uint8_t, auto... xtra>
+template<unsigned _nbits, unsigned _fbbits, typename bt = uint8_t, auto... xtra>
 class lns2b {
-	static_assert(_nbits > _rbits, "configuration not supported: not enough integer bits");
+	static_assert(_nbits > _fbbits, "configuration not supported: not enough second base bits");
 	static_assert( sizeof...(xtra) <= 1, "At most one optional extra argument is currently supported" );
-	static_assert(_nbits - _rbits < 66, "configuration not supported: the scale of this configuration is > 2^64");
-	static_assert(_rbits < 64, "configuration not supported: scaling factor is > 2^64");
+	static_assert(_nbits - _fbbits < 66, "configuration not supported: the scale of this configuration is > 2^64");
+	static_assert(_fbbits < 64, "configuration not supported: scaling factor is > 2^64");
 public:
 	typedef bt BlockType;
 
 	static constexpr unsigned nbits    = _nbits;
-	static constexpr unsigned rbits    = _rbits;
+	static constexpr unsigned fbbits   = _fbbits;
 	static constexpr Behavior behavior = {xtra...};
 
-	static constexpr double   scaling = double(1ull << rbits);
+	static constexpr double   scaling = double(1ull << fbbits);
 	static constexpr unsigned bitsInByte = 8ull;
 	static constexpr unsigned bitsInBlock = sizeof(bt) * bitsInByte;
 	static constexpr unsigned nrBlocks = (1 + ((nbits - 1) / bitsInBlock));
@@ -75,7 +75,7 @@ public:
 	static constexpr bool     SPECIAL_BITS_TOGETHER = (nbits > ((nrBlocks - 1) * bitsInBlock + 1));
 	static constexpr bt       MSU_ZERO = MSB_BIT_MASK;
 	static constexpr bt       MSU_NAN = SIGN_BIT_MASK | MSU_ZERO;  // only valid when special bits together is true
-	static constexpr int64_t  maxShift = nbits - rbits - 2;
+	static constexpr int64_t  maxShift = nbits - fbbits - 2;
 	static constexpr unsigned leftShift = (maxShift < 0) ? 0 : maxShift;
 	static constexpr int64_t  min_exponent = (maxShift > 0) ? (-(1ll << leftShift)) : 0;
 	static constexpr int64_t  max_exponent = (maxShift > 0) ? (1ll << leftShift) - 1 : 0;
@@ -422,7 +422,7 @@ public:
 	}
 	constexpr int  scale()  const noexcept {
 		ExponentBlockBinary exp(_block);
-		exp >>= rbits;
+		exp >>= fbbits;
 		return long(exp);
 	}
 	constexpr blockbinary<nbits+2, std::uint32_t, BinaryNumberType::Unsigned> fraction() const noexcept {
@@ -559,7 +559,7 @@ protected:
 				return *this = maxneg;
 			}
 			lns2b minpos(SpecificValue::minpos);
-			lns2b<nbits + 1, rbits + 1, firstBase, secondBase, bt, xtra...> halfMinpos(SpecificValue::minpos); // in log space
+			lns2b<nbits + 1, fbbits + 1, bt, xtra...> halfMinpos(SpecificValue::minpos); // in log space
[…]
