### 2023-01-12 #5058 @danielskatz
👋 @mmoelle1 - As a contributor to the software, you would normally have a COI, and in many cases, you might be listed as an author of the software/paper. Can you say anything about this case?
