### 2023-01-12 #5058 @mmoelle1
I am not an author of the software. I have started using the software some years ago and communicated actively with Theo about missing features, etc. That said, there might be a small COI as I know Theo quite well. If you can find a better reviewer please remove me from the process. Otherwise, I will be as objective as I can.
