### 2023-01-12 #5058 @danielskatz
👋 @mmoelle1 - I see a bunch of contributions from you: https://github.com/stillwater-sc/universal/commits/main?author=mmoelle1 mostly in 2018 and then a couple of minor bug fixes in 2021 and 2022.

You don't consider yourself a contributor to the software or someone who should be listed as an author of it?  I'm asking in part because of the COI issue but also because this is one of the review criteria - see https://joss.readthedocs.io/en/latest/review_checklist.html#general-checks
