### 2023-01-12 #5058 @mmoelle1
If JOSS is very strict on the definition of the term ‘contributors' (= everyone who ever created a commit to the project) then yes. Otherwise, my contribution was so minor and in an early stage of the software that I would not mind not being mentioned :-)
