### 2023-01-12 #5072 @editorialbot
[REVIEW]: Universal Numbers Library: Multi-format Variable Precision Arithmetic Library
**Submitting author:** @Ravenwater (<a href="http://orcid.org/0000-0003-0194-951X">E. Theodore L. Omtzigt</a>)
**Repository:** https://github.com/stillwater-sc/universal
**Branch with paper.md** (empty if default branch): main
**Version:** 3.68
