### 2023-01-12 58a0dd3
adding color_print and to_ieee() conversion
