+++ include/universal/number/lns2b/lns2b_impl.hpp
@@ -79,9 +79,13 @@ public:
 	static constexpr unsigned leftShift = (maxShift < 0) ? 0 : maxShift;
 	static constexpr int64_t  min_exponent = (maxShift > 0) ? (-(1ll << leftShift)) : 0;
 	static constexpr int64_t  max_exponent = (maxShift > 0) ? (1ll << leftShift) - 1 : 0;
+	static constexpr uint64_t FB_MASK = (0xFFFF'FFFF'FFFF'FFFFull >> (64 - fbbits));
+	static constexpr uint64_t SB_MASK = (0xFFFF'FFFF'FFFF'FFFFull >> (64 - (nbits - fbbits - 1)));
 
 	using BlockBinary = blockbinary<nbits, bt, BinaryNumberType::Signed>; // sign + lns2b exponent
 	using ExponentBlockBinary = blockbinary<nbits-1, bt, BinaryNumberType::Signed>;  // just the lns2b exponent
+
+	static constexpr float    base[2] = { 0.5f, 3.0f };
 
 	/// trivial constructor
 	lns2b() = default;
@@ -425,22 +429,28 @@ public:
 		exp >>= fbbits;
 		return long(exp);
 	}
-	constexpr blockbinary<nbits+2, std::uint32_t, BinaryNumberType::Unsigned> fraction() const noexcept {
-		blockbinary<nbits + 2, std::uint32_t, BinaryNumberType::Unsigned> bb{ 0 };
-		// TODO: how? and what is the size of the blockbinary? it is much bigger than nbits+2
-		return bb;
-	}
 	constexpr bool at(unsigned bitIndex) const noexcept {
 		if (bitIndex >= nbits) return false; // fail silently as no-op
 		bt word = _block[bitIndex / bitsInBlock];
 		bt mask = bt(1ull << (bitIndex % bitsInBlock));
 		return (word & mask);
 	}
-	inline constexpr bt block(unsigned b) const noexcept {
+	constexpr bt   block(unsigned b) const noexcept {
 		if (b < nrBlocks) return _block[b];
 		return bt(0); // return 0 when block index out of bounds
 	}
 
+	constexpr uint64_t extractExponent(int base) const noexcept {
+		uint64_t bits = uint64_t(_block);
+		if (base == 0) {
+			bits >>= (nbits - fbbits - 1); // normalize the value
+			bits &= FB_MASK; // null the sign bit
+		}
+		else if (base == 1) {
+			bits &= SB_MASK; // normalize the value
+		}
+		return bits;
+	}
 	explicit operator int()       const noexcept { return to_signed<int>(); }
 	explicit operator long()      const noexcept { return to_signed<long>(); }
 	explicit operator long long() const noexcept { return to_signed<long long>(); }
@@ -455,7 +465,7 @@ public:
 #endif
 
 	void debugConstexprParameters() {
-		std::cout << "constexpr parameters for " << type_tag(*this) << '\n';
+		std::cout << "constexpr parameters for " << lns2btype_tag(*this) << '\n';
 		std::cout << "scaling               " << scaling << '\n';
 		std::cout << "bitsInByte            " << bitsInByte << '\n';
 		std::cout << "bitsInBlock           " << bitsInBlock << '\n';
@@ -470,6 +480,12 @@ public:
 		std::cout << "BLOCK_MSB_MASK        " << to_binary(BLOCK_MSB_MASK, bitsInBlock) << '\n';
 		std::cout << "MSU_ZERO              " << to_binary(MSU_ZERO, bitsInBlock) << '\n';
 		std::cout << "MSU_NAN               " << to_binary(MSU_NAN, bitsInBlock) << '\n';
+		std::cout << "maxShift              " << maxShift << '\n';
[…]
