### 2023-01-12 #5072 @Ravenwater
@danielskatz do I need to worry about this? What is prohibiting gitinspector from succeeding?
