### 2023-01-12 #5072 @mmoelle1
My apologize but where should I run the command

I never reviewed for JOSS before.
