### 2023-01-12 #5072 @mmoelle1
> I think you need to do this in directly in the GitHub web interface in [this issue](https://github.com/openjournals/joss-reviews/issues/5072). (I don't think this will work by email but you can try it if you want.) An editorialbot command should be the first thing in a new comment in an issue. And I'm fairly sure you will need to check items in the checklist in the GitHub web interface as well.

Got it and it works.
