### 2023-01-12 4cede49
bug fix on negative lns2b encodings to value conversion
