### 2023-01-12 5d1ef06
adding traits to lns and lns2s common functions, testing it with lns_comparison
