+++ include/universal/number/lns/manipulators.hpp
@@ -14,51 +14,45 @@
 namespace sw { namespace universal {
 
 	// Generate a type tag for this lns
-	template<unsigned nbits, unsigned rbits, typename BlockType, auto... xtra>
-	inline std::string type_tag(const lns<nbits, rbits, BlockType, xtra...>& = {}) {
+	template<typename LnsType,
+		std::enable_if_t< is_lns<LnsType>, bool> = true
+	>
+	inline std::string type_tag(const LnsType & = {}) {
 		std::stringstream s;
 		s << "lns<"
-			<< std::setw(3) << nbits << ", "
-			<< std::setw(3) << rbits << ", "
-			<< std::setw(10) << type_tag(Behavior{xtra...}) << ", "
-			<< typeid(BlockType).name() << '>';
+			<< std::setw(3) << LnsType::nbits << ", "
+			<< std::setw(3) << LnsType::rbits << ", "
+			<< typeid(LnsType::BlockType).name() << ", "
+		<< std::setw(10) << type_tag(Behavior{ LnsType::behavior }) << '>';
 		return s.str();
 	}
 
-	// report dynamic range of a type, specialized for lns
-	template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
-	inline std::string dynamic_range(const lns<nbits, rbits, bt, xtra...>& a) {
+	template<typename LnsType,
+		std::enable_if_t< is_lns<LnsType>, bool> = true
+	>
+	inline std::string range(const LnsType & = {}) {
 		std::stringstream s;
-		lns<nbits, rbits, bt, xtra...> b(SpecificValue::maxneg), c(SpecificValue::minneg), d(SpecificValue::minpos), e(SpecificValue::maxpos);
-		s << type_tag(a) << ": ";
-		s << "minpos scale " << std::setw(10) << d.scale() << "     ";
-		s << "maxpos scale " << std::setw(10) << e.scale() << '\n';
-		s << "[" << b << " ... " << c << ", 0, " << d << " ... " << e << "]\n";
-		s << "[" << to_binary(b) << " ... " << to_binary(c) << ", 0, " << to_binary(d) << " ... " << to_binary(e) << "]\n";
-		return s.str();
-	}
-
-	template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
-	inline std::string range() {
-		std::stringstream s;
-		lns<nbits, rbits, bt, xtra...> b(SpecificValue::maxneg), c(SpecificValue::minneg), d(SpecificValue::minpos), e(SpecificValue::maxpos);
+		LnsType b(SpecificValue::maxneg), c(SpecificValue::minneg), d(SpecificValue::minpos), e(SpecificValue::maxpos);
 		s << "[" << b << " ... " << c << ", 0, " << d << " ... " << e << "]\n";
 		return s.str();
 	}
 
 	// report if a native floating-point value is within the dynamic range of the lns configuration
-	template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
+	template<typename LnsType,
+		std::enable_if_t< is_lns<LnsType>, bool> = true
+	>
 	inline bool isInRange(double v) {
-		using LNS = lns<nbits, rbits, bt, xtra...>;
-		LNS a{};
+		LnsType a{};
 
-		bool inRange = true;
-		if (v > double(a.maxpos()) || v < double(a.maxneg())) inRange = false;
-		return inRange;
+		bool inside = true;
+		if (v > double(a.maxpos()) || v < double(a.maxneg())) inside = false;
+		return inside;
 	}
 
[…]
