### 2023-01-12 da0c982
adding typename to BlockType reference to fix gcc compile error
