### 2023-01-14 eca9e98
bug fix SpecificValue patterns for lns2b
