### 2023-01-14 fa92661
type_tag alignement with other Universal number systems
