+++ validation/hw/alu/fp8_4_subsupnosat.cpp
@@ -1,4 +1,4 @@
-// bfloat16.cpp: testbench for a bfloat16 hardware ALU
+// fp8_4_subsupnosat.cpp: testbench for a fp<8,4, subnormals, supernormals, non-saturating> hardware ALU
 //
 // Copyright (C) 2017-2023 Stillwater Supercomputing, Inc.
 //
@@ -8,24 +8,28 @@
 #include <universal/hw/alu.hpp>
 #include <universal/number/cfloat/cfloat.hpp>
 
-
 int main(int argc, char** argv)
 try {
 	using namespace sw::universal;
 
+	// fp8_4_subsupnosat has subnormals and supernormals
+	constexpr unsigned nbits = 8;
+	constexpr unsigned es = 4;
+	using fp8_4_subsupnosat = cfloat<nbits, es, std::uint16_t, true, true, false>;
+
+	using Real = fp8_4_subsupnosat;
+
+	EnumerateValidEncodings<Real>(std::cout);
+
+	return 0;
+
 	if (!(argc == 2 || argc == 4)) {
-		std::cerr << "Usage: hw_bfloat16 [add | sub | mul | div | sqrt] [a b]\n";
-		std::cerr << "Example: hw_bfloat16 add 1.5 -1.5\n";
+		std::cerr << "Usage: hw_fp8_4_subsupnosat [add | sub | mul | div | sqrt] [a b]\n";
+		std::cerr << "Example: hw_fp8_4_subsupnosat add 1.5 -1.5\n";
 		return EXIT_SUCCESS;  // needed for regression test success
 	}
 	std::string op = argv[1];
-	std::cout << "generating bfloat16 test vectors for " << op;
-	// bfloat16 does not have subnormals
-	constexpr unsigned nbits = 16;
-	constexpr unsigned es = 8;
-	using bfloat16 = cfloat<nbits, es, std::uint16_t, false, false, false>;
-
-	using Real = bfloat16;
+	std::cout << "generating fp<8,4,sub, sup, nosat> test vectors for " << op;
 
 	float fa, fb;
 	if (argc == 4) {
