### 2023-01-14 5cc0887
streamlining new HW test vector generators
