+++ validation/hw/alu/cfloat.cpp
@@ -1,6 +1,6 @@
-// cfloat.cpp: testbench for a classic float hardware ALU
+// cfloat.cpp: test vector generator for a classic float hardware ALU
 //
-// Copyright (C) 2017-2021 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017-2023 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #include <iostream>
@@ -14,51 +14,37 @@ int main(int argc, char** argv)
 try {
 	using namespace sw::universal;
 
+	// fp<8,2> without subnormals, supernormals or saturation
+	constexpr unsigned nbits = 8;
+	constexpr unsigned es = 2;
+	using fp8_2_nosubsupsat = cfloat<nbits, es, std::uint16_t, false, false, false>;
+
+	using Real = fp8_2_nosubsupsat;
+
 	if (argc != 4) {
 		std::cerr << "Usage: hw_cfloat [add | sub | mul | div | sqrt] a b\n";
 		std::cerr << "Example: hw_cfloat add 1.5 -1.5\n";
 		return EXIT_SUCCESS;  // needed for regression test success
 	}
 	std::string op = argv[1];
-	float fa = atof(argv[2]);
-	float fb = atof(argv[3]);
-	std::cout << op << " " << fa << " and " << fb << '\n';
+	std::cout << "generating fp<8,4,nosub,nosup,nosat> test vectors for " << op;
 
-	cfloat<8, 2> a, b, c;
-	a = fa;
-	b = fb;
-
-	// decode the operation
-	ALU_OPS alu_op;
-	if (op == "add") {
-		alu_op = ALU_OPS::ADD;
-		c = ArithmeticLogicUnit(alu_op, a, b);
-		std::cout << a << " + " << b << " = " << c << '\n';
-		std::cout << to_binary(a, true) << " + " << to_binary(b, true) << " = " << to_binary(c, true) << '\n';
+	float fa, fb;
+	if (argc == 4) {
+		fa = atof(argv[2]);
+		fb = atof(argv[3]);
+		std::cout << " " << fa << ' ' << fb << '\n';
+		ExecuteOp<Real>(op, fa, fb);
 	}
-	else if (op == "sub") {
-		alu_op = ALU_OPS::SUB;
-		c = ArithmeticLogicUnit(alu_op, a, b);
-		std::cout << a << " - " << b << " = " << c << '\n';
-		std::cout << to_binary(a, true) << " - " << to_binary(b, true) << " = " << to_binary(c, true) << '\n';
-	}
-	else if (op == "mul") {
-		alu_op = ALU_OPS::MUL;
-		c = ArithmeticLogicUnit(alu_op, a, b);
-		std::cout << a << " * " << b << " = " << c << '\n';
-		std::cout << to_binary(a, true) << " * " << to_binary(b, true) << " = " << to_binary(c, true) << '\n';
-	}
-	else if (op == "div") {
-		alu_op = ALU_OPS::DIV;
-		c = ArithmeticLogicUnit(alu_op, a, b);
-		std::cout << a << " / " << b << " = " << c << '\n';
-		std::cout << to_binary(a, true) << " / " << to_binary(b, true) << " = " << to_binary(c, true) << '\n';
-	}
-	else if (op == "sqrt") {
-		alu_op = ALU_OPS::SQRT;
-		c = ArithmeticLogicUnit(alu_op, a, b);
-		std::cout << "sqrt(" << a << ") = " << c << '\n';
-		std::cout << "sqrt(" << to_binary(a, true) << " = " << to_binary(c, true) << '\n';
+	else {
[…]
