### 2023-01-14 cf8bd9f
adding new minpos/maxpos range comparison among 8-bit number systems
