+++ validation/hw/alu/ranges.cpp
@@ -1,0 +1,145 @@
+// ranges.cpp: comparing minpos/maxpos ranges for different small encoding number systems
+//
+// Copyright (C) 2017-2023 Stillwater Supercomputing, Inc.
+//
+// This file is part of the universal numbers project, which is released under an MIT Open Source license.
+#include <iostream>
+#include <iomanip>
+#include <fstream>
+#include <string>
+#include <universal/hw/alu.hpp>
+#include <universal/number/lns/lns.hpp>
+#include <universal/number/lns2b/lns2b.hpp>
+#include <universal/number/cfloat/cfloat.hpp>
+#include <universal/number/posit/posit.hpp>
+
+static constexpr unsigned TYPENAMEWIDTH = 80;
+static constexpr unsigned COLWIDTH = 15;
+
+void PrintMinposMaxposHeader(std::ostream& ostr, bool tsv) {
+	if (tsv) {
+		ostr << "number type\tminpos\tmaxpos\n";
+	}
+	else {
+		ostr << std::setw(TYPENAMEWIDTH) << "number type" << std::setw(COLWIDTH) << "minpos" << std::setw(COLWIDTH) << "maxpos" << '\n';
+	}
+}
+
+template<typename NumberType>
+void PrintMinposMaxpos(std::ostream& ostr, bool tsv) {
+	using namespace sw::universal;
+	if (tsv) {
+		ostr << type_tag(NumberType()) << "\t" << NumberType(SpecificValue::minpos) << "\t" << NumberType(SpecificValue::maxpos) << '\n';
+	}
+	else {
+		ostr << std::setw(TYPENAMEWIDTH) << type_tag(NumberType()) << std::setw(COLWIDTH) << NumberType(SpecificValue::minpos) << std::setw(COLWIDTH) << NumberType(SpecificValue::maxpos) << '\n';
+	}
+}
+
+int main(int argc, char** argv)
+try {
+	using namespace sw::universal;
+
+	using fp8_3_nosubsup = cfloat<8,3, std::uint8_t, false, false, false>;
+	using fp8_4_nosubsup = cfloat<8,4, std::uint8_t, false, false, false>;
+	using fp8_5_nosubsup = cfloat<8,5, std::uint8_t, false, false, false>;
+	
+	using fp8_3_subsup = cfloat<8,3,   std::uint8_t, true, true, false>;
+	using fp8_4_subsup = cfloat<8,4,   std::uint8_t, true, true, false>;
+	using fp8_5_subsup = cfloat<8,5,   std::uint8_t, true, true, false>;
+
+	using lns8_3 = lns<8,3>;
+	using lns8_4 = lns<8,4>;
+	using lns8_5 = lns<8,5>;
+
+	using lns2b8_3 = lns2b<8,3>;
+	using lns2b8_4 = lns2b<8,4>;
+	using lns2b8_5 = lns2b<8,5>;
+
+	using posit8_0 = posit<8,0>;
+	using posit8_1 = posit<8,1>;
+	using posit8_2 = posit<8,2>;
+
+	bool tsv = false;
+	std::string filename("range.tsv");
+	if (argc > 2) {
+		std::cerr << "Usage: hw_range [tab-separated file name]\n";
+		return EXIT_SUCCESS;
+	}
+	else if (argc == 2) {
+		tsv = true;
[…]
