### 2023-01-14 e88a6ab
redesign encoding enumeration to depend on operator++()
