### 2023-01-14 159be60
enumeration via operator++() now honors configured subnormal range
