### 2023-01-14 9e2d254
bug fix for assignment now that subnormals are assumed to not exist in cfloat without subnormals
