### 2023-01-14 2b81814
bug fix for decrement now that subnormals are assumed to not exist in cfloat without subnormals
