+++ static/cfloat/conversion/assignment.cpp
@@ -1,6 +1,6 @@
 // assignment.cpp: functional tests for assignments of native types to cfloats
 //
-// Copyright (C) 2017-2022 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017-2023 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #include <universal/utility/directives.hpp>
