### 2023-01-15 3c3a1f4
documenting the subtle difference between iszero and iszeroencoding in encodings that do/do not support subnormals
