+++ include/universal/verification/cfloat_test_suite.hpp
@@ -777,11 +777,11 @@ namespace sw { namespace universal {
 					fractionBits = 2 * fbits;
 					integerSet = 4;
 				}
-				size_t NR_VALUES = (1ull << fractionBits);
+				size_t NR_ENCODINGS = (1ull << fractionBits);
 				b.setscale(scale);
 				for (size_t i = 1; i < integerSet; ++i) {  // 01, 10, 11.fffff: state 00 is not part of the encoding as that would represent a denormal
 					size_t integerBits = (i << abits);
-					for (size_t f = 0; f < NR_VALUES; ++f) {
+					for (size_t f = 0; f < NR_ENCODINGS; ++f) {
 						size_t btbits = integerBits | (f << rbits);
 						b.setbits(btbits);
 //						b.setbits(integerBits + f);
@@ -892,7 +892,7 @@ namespace sw { namespace universal {
 		constexpr bool isSaturating = CfloatConfiguration::isSaturating;
 
 		int nrOfTestFailures{ 0 };
-		constexpr size_t NR_VALUES = (1u << nbits);
+		constexpr size_t NR_ENCODINGS = (1u << nbits);
 		cfloat<nbits, es, bt, hasSubnormals, hasSupernormals, isSaturating> a{};
 
 		// ADD
@@ -900,7 +900,7 @@ namespace sw { namespace universal {
 			//constexpr size_t abits = CfloatConfiguration::abits;
 			constexpr size_t fbits = CfloatConfiguration::fbits;
 			blocktriple<fbits, op, bt> b;   // the size of the blocktriple is configured by the number of fraction bits of the source number system
-			for (size_t i = 0; i < NR_VALUES; ++i) {
+			for (size_t i = 0; i < NR_ENCODINGS; ++i) {
 				a.setbits(i);
 				a.normalizeAddition(b);
 				if (double(a) != double(b)) {
@@ -923,7 +923,7 @@ namespace sw { namespace universal {
 			constexpr size_t fbits = CfloatConfiguration::fbits;
 			blocktriple<fbits, op, bt> b;   // the size of the blocktriple is configured by the number of fraction bits of the source number system
 			blocktriple<2 * fbits, BlockTripleOperator::REP, bt> ref;
-			for (size_t i = 0; i < NR_VALUES; ++i) {
+			for (size_t i = 0; i < NR_ENCODINGS; ++i) {
 				a.setbits(i);
 				a.normalizeMultiplication(b);
 				ref = double(b);
@@ -948,7 +948,7 @@ namespace sw { namespace universal {
 			constexpr size_t fbits = CfloatConfiguration::fbits;
 			blocktriple<fbits, op, bt> b;   // the size of the blocktriple is configured by the number of fraction bits of the source number system
 			blocktriple<2 * fbits, BlockTripleOperator::REP, bt> ref;
-			for (size_t i = 0; i < NR_VALUES; ++i) {
+			for (size_t i = 0; i < NR_ENCODINGS; ++i) {
 				a.setbits(i);
 				a.normalizeDivision(b);
 				ref = double(b);
@@ -1255,7 +1255,7 @@ namespace sw { namespace universal {
 		constexpr bool isSaturating    = TestType::isSaturating;
 		using Cfloat = sw::universal::cfloat<nbits, es, BlockType, hasSubnormals, hasSupernormals, isSaturating>;
 
-		constexpr size_t NR_VALUES = (size_t(1) << nbits);
+		constexpr size_t NR_ENCODINGS = (size_t(1) << nbits);
 		int nrOfFailedTests = 0;
 
[…]
