### 2023-01-16 29b8ba7
moving the performance runner into its own benchmark category
