+++ benchmark/performance/arithmetic/integer/performance.cpp
@@ -14,7 +14,7 @@
 // is representable
 #include <universal/functions/isrepresentable.hpp>
 #include <universal/verification/test_status.hpp> // ReportTestResult
-#include <universal/verification/performance_runner.hpp>
+#include <universal/benchmark/performance_runner.hpp>
 
 /*
    The goal of the arbitrary integers is to provide a constrained big integer type
