### 2023-01-17 3a23290
adding attribute tests for 2-base lns
