### 2023-01-18 0b7463b
adding attributes.hpp to lns2b
