+++ include/universal/number/lns2b/lns2b.hpp
@@ -47,7 +47,7 @@
 
 // useful functions to work with logarithmic numbers
 #include <universal/number/lns2b/manipulators.hpp>
-//#include <universal/number/lns2b/attributes.hpp>
+#include <universal/number/lns2b/attributes.hpp>
 
 ///////////////////////////////////////////////////////////////////////////////////////
 /// math functions
