### 2023-01-18 eea6287
adding number traits to lns2b library
