+++ include/universal/number/lns2b/lns2b.hpp
@@ -24,18 +24,24 @@
 
 ////////////////////////////////////////////////////////////////////////////////////////
 // enable/disable the ability to use literals in binary logic and arithmetic operators
-#if !defined(LNS_ENABLE_LITERALS)
+#if !defined(LNS2B_ENABLE_LITERALS)
 // default is to enable them
-#define LNS_ENABLE_LITERALS 1
+#define LNS2B_ENABLE_LITERALS 1
 #endif
 
 ////////////////////////////////////////////////////////////////////////////////////////
 // enable throwing specific exceptions for logarithmic number system arithmetic errors
 // left to application to enable
-#if !defined(LNS_THROW_ARITHMETIC_EXCEPTION)
+#if !defined(LNS2B_THROW_ARITHMETIC_EXCEPTION)
 // default is to use std::cerr for signalling an error
-#define LNS_THROW_ARITHMETIC_EXCEPTION 0
+#define LNS2B_THROW_ARITHMETIC_EXCEPTION 0
 #endif
+
+///////////////////////////////////////////////////////////////////////////////////////
+// bring in the trait functions
+#include <universal/traits/number_traits.hpp>
+#include <universal/traits/arithmetic_traits.hpp>
+#include <universal/common/number_traits_reports.hpp>
 
 ////////////////////////////////////////////////////////////////////////////////////////
 /// INCLUDE FILES that make up the library
