+++ static/lns2b/api/attributes.cpp
@@ -12,7 +12,6 @@
 #define LNS2B_THROW_ARITHMETIC_EXCEPTION 1
 // third: enable support for native literals in logic and arithmetic operations
 #define LNS2B_ENABLE_LITERALS 1
-// minimum set of include files to reflect source code dependencies
 #include <universal/number/lns2b/lns2b.hpp>
 #include <universal/verification/test_reporters.hpp>
 
@@ -32,25 +31,26 @@ try {
 
 	{
 		std::cout << "\nDynamic ranges of 2-base logarithmic floating-point arithmetic types\n";
-		std::cout << dynamic_range< lns2b<  8, 6> >() << '\n';
-		std::cout << dynamic_range< lns2b< 16,10> >() << '\n';
-		std::cout << dynamic_range< lns2b< 32,22> >() << '\n';
+		std::cout << dynamic_range< lns2b<  8, 5> >() << '\n';    //  8 - 1 -  5 = 2 scale exponent bits: 3^##
+		std::cout << dynamic_range< lns2b< 16,12> >() << '\n';    // 16 - 1 - 12 = 3 scale exponent bits: 3^###
+		std::cout << dynamic_range< lns2b< 24,19> >() << '\n';    // 24 - 1 - 19 = 4 scale exponent bits: 3^####
+		std::cout << dynamic_range< lns2b< 32,26> >() << '\n';    // 32 - 1 - 26 = 5 scale exponent bits: 3^#####
 	}
 
 	{
 		std::cout << "\nMinmax of 2-base logarithmic floating-point\n";
-		std::cout << minmax_range< lns2b< 8, 6> >() << '\n';
-		std::cout << minmax_range< lns2b<16,10> >() << '\n';
-		std::cout << minmax_range< lns2b<24,16> >() << '\n';
-		std::cout << minmax_range< lns2b<32,22> >() << '\n';
+		std::cout << minmax_range< lns2b< 8, 5> >() << '\n';
+		std::cout << minmax_range< lns2b<16,12> >() << '\n';
+		std::cout << minmax_range< lns2b<24,19> >() << '\n';
+		std::cout << minmax_range< lns2b<32,26> >() << '\n';
 	}
 	
 	{
 		std::cout << "\nValue ranges of 2-base logarithmic floating-point\n";
-		std::cout << symmetry_range< lns2b< 8, 6> >() << '\n';
-		std::cout << symmetry_range< lns2b<16,10> >() << '\n';
-		std::cout << symmetry_range< lns2b<24,16> >() << '\n';
-		std::cout << symmetry_range< lns2b<32,22> >() << '\n';
+		std::cout << symmetry_range< lns2b< 8, 5> >() << '\n';
+		std::cout << symmetry_range< lns2b<16,12> >() << '\n';
+		std::cout << symmetry_range< lns2b<24,19> >() << '\n';
+		std::cout << symmetry_range< lns2b<32,26> >() << '\n';
 	}
 
 	{
@@ -68,12 +68,13 @@ try {
 
 	{
 		std::cout << "\nNumber traits\n";
-//		numberTraits< lns2b<32, 16> >(std::cout);
+		numberTraits< lns2b<8, 4> >(std::cout);
 	}
 
 	{
 		std::cout << "\nComparitive Number traits\n";
-//		compareNumberTraits< lns2b<24, 12>, lns2b<32, 16> >(std::cout);
+		compareNumberTraits< lns2b<10, 6>, lns2b<12, 7> >(std::cout);
+		threeWayCompareNumberTraits< float, lns2b<10, 6>, lns2b<12, 7> >(std::cout);
 	}
 
 	ReportTestSuiteResults(test_suite, nrOfFailedTestCases);
[…]
