### 2023-01-18 2ad83cb
adding 2-base lns polynomial evaluation
