### 2023-01-20 6e3f623
Update README.md
Set C++ standard in cmake instructions appropriately
