+++ README.md
@@ -189,6 +189,7 @@ project("my-numerical-experiment")
 find_package(UNIVERSAL CONFIG REQUIRED)
 
 add_executable(${PROJECT_NAME} src/mymain.cpp)
+set_property(TARGET ${PROJECT_NAME} PROPERTY CXX_STANDARD 17)
 target_link_libraries(${PROJECT_NAME} universal::universal)
 ```
 
