### 2023-01-22 9d54ab8
Fix cmake instructions (#321)
* Fix capitalization of universal dependency in cmake instructions (README.md)

Fix capitalization of universal dependency in cmake instructions (README.md)

* Update README.md

Set C++ standard in cmake instructions appropriately
