+++ README.md
@@ -189,7 +189,8 @@ project("my-numerical-experiment")
 find_package(UNIVERSAL CONFIG REQUIRED)
 
 add_executable(${PROJECT_NAME} src/mymain.cpp)
-target_link_libraries(${PROJECT_NAME} UNIVERSAL::UNIVERSAL)
+set_property(TARGET ${PROJECT_NAME} PROPERTY CXX_STANDARD 17)
+target_link_libraries(${PROJECT_NAME} universal::universal)
 ```
 
 ## Controlling the build to include different components
