### 2023-01-24 2a89a39
V3.67 (#322)
* bumping SEMVER to v3.67

* moving CI to new dev branch

* disabling the JOSS darft paper workflow

* adding JOSS submission to the README

* updating bibtex with DOIs

* changing mdlns to lns2b to simplify template API

* moving mdlns hierarchy to lns2b

* renaming files using new lns2b type name

* renaming mdlns to lns2b

* WIP: stitching back together the nw lns2b type

* adding color_print and to_ieee() conversion

* bug fix on negative lns2b encodings to value conversion

* adding traits to lns and lns2s common functions, testing it with lns_comparison

* adding typename to
[…]
