### 2023-01-24 ed5e60a
streamlining build, test, dev, and release containers
