+++ build_test_container.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-
-docker build --force-rm -f Dockerfile.test -t stillwater/alu:latest .
-
