### 2023-01-25 #5072 @Ravenwater
I just release v3.67. It adds a 2-base logarithmic number system to Universal with some comparisons between small cfloats, lns, and lns2b, including the different proposed IEEE-754 FP8 formats for Deep Learning.
