### 2023-01-26 #5072 @yboettcher
A few notes from my side, some of which are most likely fine the way they are, but I would still like to clarify these.

The first is in regards to the license and Contributing points. The license states that Stillwater Supercomputing holds the copyright to universal. I would assume that if someone were to make a code contribution to universal, this would mean that they transfer any (copy)rights they have to Stillwater. I am not an expert on legal things, but does the existence of a LICENSE file with a copyright holder imply this (in which case I would check off these items), or would it be better to explicitely mention this either in the Readme or a separate Contribution.md file?

The second one is about the authors: like previous publications on this library, the paper lists Theodore Omtzigt and James Quinlan as authors. However, judging the Contributors.md file, there are some more people who have worked on this. I would assume (and like to clarify) that authors refers to the "owners/driving forces" behind the software project, in which case the current list of authors looks sensible to me.

Some nitpicks on the software paper:

In the example at the end of the summary, the "color coded" form is not color coded as it is in the terminal. Seeing that the paper is generated from markdown and it doesn't seem like markdown supports coloring easily, I guess there is nothing to do about this.

The sentences on line 37 and 38 seem a bit strange to me. On l.37, the "environmental impacts and financial costs" are attributed to a single cause. On l.38 it says "these both result from". Due to the use of the plural, it appears to be referencing the "environmental impacts and financial costs" again and attributes them to a different cause. Continuing on l.38, I'm a bit confused about the "growth". I feel like this should say "growing requirements".

On line 43, there appears to be a broken reference (kharya:2020).

In both the paper and the github readme, you state that several projects "leverage" universal/posits. I feel like this is a little suboptimal wording. Out of the mentioned projects, I could only find references to universal in G+SMO, whereas "leverage" suggests (to me) that there is some integration of universal (or even a dependency on universal) into these projects, which I feel is not the case for mtl4, odeint and bembel.

And two notes on the readme:

Right at the beginning of the "Motivation" section, "the 64-bit format" is referenced, which might be a little ambiguous in the context of a library implementing many formats.

The `make posit_32bit_posit` command in the "Verification Suite" section does not work (for me atleast, no rule to for posit_32bit_posit). I suspect that this might just need updating of the target name.
