### 2023-01-27 #5072 @jamesquinlan
@yboettcher thank you for the feedback.  I will defer to @Ravenwater for many of these points, and I'll address what I think I can.

In the example, the color came through in markdown using ```cpp ```, but the output from the terminal is just plain markdown.  But I note that in the same example, "some" colors, like a red left parenthesis and an olive green negative number.  I'm not sure how to fix that (or if it can be).

L38+39.  I agree.  What we were trying to say comes from haidar et al. 2018 "...the increasing performance of HPC systems comes with a cost: electrical power consumption. This leads to the following two main concerns, an increase of the power bills beyond affordable budgets and an increasing impact on the environment."  This is directly related to processing (and storage).   I will rewrite that thought better.    I also agree, "growing requirements" is more clear.

It looks like kharya:2020 is missing the @ in front to form the link to reference.

Thanks again for the review and feedback.
