### 2023-01-27 #5072 @Ravenwater
@yboettcher The copyright holder is the person or entity that can claim the right to have created a piece of work, or to whom such right has been transferred. The MIT license is not about transferring copyright, but about granting the legal right to use the (copyrighted) work as stated in the license. If you write a class or function or application, you have the copyright to it, even if its actual use depends on third-party work, such as the Universal build and test system, if you include it into the Universal repo.

Once you add your contribution to the software package and distribute the whole package, the package has multiple copyright holders: the original authors, and you. You are allowed to distribute the parts from the original package because the MIT license allows you to "use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software".
