### 2023-01-27 #5072 @Ravenwater
@yboettcher @jamesquinlan we need to make it clearer that if you want to build a specific target, then you would need to have enabled that build target in the build system. So if you want to build a posit target, such as `make posit_32bit_posit`, you would need to have executed a `cmake -DBUILD_NUMBER_POSITS=ON ..` to enable that build target.
