### 2023-01-27 #5072 @Ravenwater
@yboettcher  In the ArXiv paper: https://arxiv.org/ftp/arxiv/papers/2012/2012.11011.pdf we present mixed-precision results using Universal to extend ODEint, and the fused dot product is being enabled automatically in MTL4 when a posit scalar type is detected. And @yboettcher leveraged the posits of Universal to extend Bembel to generate new results.

So in all three cases, these platforms, which were written to be specialized with libraries like Universal, _leverage_ Universal to create new results.
