### 2023-01-28 #5072 @jamesquinlan
@Ravenwater, I agree.  We can add that to the article's body and the readme file.  I would also like to add a sentence on the new additions of v3.67 in the Summary.

As for the sentence in the statement of need,  a slight modification should clarify: ``High-performance computing (HPC), machine learning, and deep learning tasks [e.g., @carmichael:2019; @cococcioni2022small;@desrentes:2022posit8] have increased environmental impacts and financial costs due to massive energy consumption [@haidar:2018b] from the growing processing requirements.
