### 2023-01-29 33a8aa0
removing the static_assert for long double being an extended precision floating-point
