### 2023-01-29 ea23361
removing CodeShip CI as they are no longer supporting open-source projects
