+++ codeship-steps.yml
@@ -1,6 +1,4 @@
-- name: log_env
-  service: universal-test-suite
-  command: printenv
-- name: test_runner
-  service: universal-test-suite
-  command: make test
+universal-test-suite:
+  build: .
+  environment:
+    TEST_VAR: today
