+++ joss/paper.md
@@ -94,7 +94,7 @@ color coded  : 0b0.0111'1111.001'1110'0000'0110'0101'0010
 
 # Statement of need
 
-High-performance computing (HPC), machine learning, and deep learning tasks [e.g., @carmichael:2019; @cococcioni2022small;@desrentes:2022posit8] have increased environmental impacts and financial costs due to massive energy consumption [@haidar:2018b].  These both result from growth requirements in processing and storage.  In addition to redesigning algorithms to minimize data movement and processing,  modern systems increasingly support multi-precision arithmetic in hardware [@haidar:2018a].  Recently, NVIDIA added support for low-precision formats to its top-level GPUs to perform tensor operations [@choquette2021nvidia], including a 19-bit format having an exponent of 8 bits and a mantissa of 10 bits (see also [@intel:2018; kharya:2020].  In addition, the "Brain Floating Point Format," commonly referred to as "bfloat16," is a format developed by Google that enables the training and operation of deep neural networks using specialized processors called Tensor Processing Units, or TPUs, at higher performance and cheaper cost [@wang2019bfloat16].  As a result, we see a trend to redesign many standard algorithms.  In particular, designing fast and energy-efficient linear solvers is an active area of research where low-precision numerics plays a fundamental role [@carson:2018; @haidar:2017; @haidar:2018a; @haidar:2018b; @higham:2019].
+High-performance computing (HPC), machine learning, and deep learning tasks [e.g., @carmichael:2019; @cococcioni2022small;@desrentes:2022posit8] have increased environmental impacts and financial costs due to massive energy consumption [@haidar:2018b].  These both result from growth requirements in processing and storage.  In addition to redesigning algorithms to minimize data movement and processing,  modern systems increasingly support multi-precision arithmetic in hardware [@haidar:2018a].  Recently, NVIDIA added support for low-precision formats to its top-level GPUs to perform tensor operations [@choquette2021nvidia], including a 19-bit format having an exponent of 8 bits and a mantissa of 10 bits (see also [@intel:2018; @kharya:2020].  In addition, the "Brain Floating Point Format," commonly referred to as "bfloat16," is a format developed by Google that enables the training and operation of deep neural networks using specialized processors called Tensor Processing Units, or TPUs, at higher performance and cheaper cost [@wang2019bfloat16].  As a result, we see a trend to redesign many standard algorithms.  In particular, designing fast and energy-efficient linear solvers is an active area of research where low-precision numerics plays a fundamental role [@carson:2018; @haidar:2017; @haidar:2018a; @haidar:2018b; @higham:2019].
 
 
 While the primary motivation for low-precision arithmetic is its high performance and energy efficiency, mixed-precision algorithm designs aim to identify and exploit opportunities to right-scale the number of systems used for critical computational paths representing the execution bottleneck. Furthermore, when these algorithms are incorporated into embedded devices and custom hardware engines, we approach optimal performance and power efficiency. Therefore, investigations into computational mathematics and measuring mixed-precision algorithms' accuracy, efficiency, robustness, and stability are needed.
