### 2023-02-06 #5072 @yboettcher
@Ravenwater Thanks for clarifying the copyright thing.

As for the `make posit_32bit_posit`, the following chain of commands returns an error (`make: *** No rule to make target 'posit_32bit_posit'.  Stop.`):

```bash
git clone https://github.com/stillwater-sc/universal
cd universal/
mkdir build
cd build
cmake -DBUILD_NUMBER_POSITS=ON ..
make posit_32bit_posit
```

I can run `make posit` just fine though, so I guess `posit_32bit_posit` is an outdated target name.

It also seems like the list of provided number systems in the readme (section "Structure of the tree") might need updating.
Out of those, the adaptiveint/adaptivefloat/adaptiveposit don't exist in include/universal.
The quire only contains an `exception.hpp`, the unum2 does not contain a "main" `unum2.hpp` and the float also seems to be missing content.
On the other hand, you mention SORNs in the software paper, but not in the readme.

I also think it would be helpful if there was some way of knowing how to use the different number system templates. You have an example that shows how that works for posits, but as for the lns numbers (for example), I would not know, how many template parameters there are, or what they mean.
