### 2023-02-06 523457c
cody hygiene for edge case es == 0
