+++ tools/cmd/plimits.cpp
@@ -1,8 +1,9 @@
 // plimits.cpp: cli to show the numeric_limits<> of the standard posits
 //
-// Copyright (C) 2017-2022 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017-2023 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
+#include <universal/utility/directives.hpp>
 #include <universal/number/posit/posit.hpp>
 
 template<size_t nbits, size_t es>
