### 2023-02-06 62faadd
fixing up README using JOSS reviewer feedback
