+++ README.md
@@ -107,14 +107,14 @@ For Ubuntu, snap will install the latest cmake, and would be the preferred metho
 > sudo snap install cmake --classic
 ```
 
-The Universal Library is a pure C++ template library without any further dependencies, even for the regression test suites,
-to enable hassle-free installation and use.
+The Universal Library is a pure C++ template library without any further dependencies, 
+even for the regression test suites, to enable hassle-free installation and use.
 
 Clone the GitHub repo, and you are ready to build the different components of the Universal library.  
 The library contains tools for using integers, decimals, fixed-points, floats, posits, valids, and logarithmic
 number systems. It includes educational programs that showcase simple use cases to familiarize yourself with
 different number systems and application examples to highlight the use of other number systems to gain performance
-or numerical accuracy. Finally, each number system offers its verification suite. 
+or numerical accuracy. Finally, each number system offers its own verification suite. 
 
 The easiest way to become familiar with all the options in the build process is to fire up the CMake GUI (or ccmake if you are on a headless server). The CMake output will summarize which options have been set.  
 The output will look something like this:
@@ -125,13 +125,170 @@ $ cd universal
 $ mkdir build
 $ cd build
 $ cmake ..
+
+ _____  _____  ____  _____  _____  ____   ____  ________  _______     ______        _       _____
+|_   _||_   _||_   \|_   _||_   _||_  _| |_  _||_   __  ||_   __ \  .' ____ \      / \     |_   _|
+  | |    | |    |   \ | |    | |    \ \   / /    | |_ \_|  | |__) | | (___ \_|    / _ \      | |
+  | '    ' |    | |\ \| |    | |     \ \ / /     |  _| _   |  __ /   _.____`.    / ___ \     | |   _
+   \ \__/ /    _| |_\   |_  _| |_     \ ' /     _| |__/ | _| |  \ \_| \____) | _/ /   \ \_  _| |__/ |
+    `.__.'    |_____|\____||_____|     \_/     |________||____| |___|\______.'|____| |____||________|
+
+-- The C compiler identification is GNU 9.4.0
+-- The CXX compiler identification is GNU 9.4.0
+-- The ASM compiler identification is GNU
+-- Found assembler: /usr/bin/cc
+-- Detecting C compiler ABI info
+-- Detecting C compiler ABI info - done
+-- Check for working C compiler: /usr/bin/cc - skipped
+-- Detecting C compile features
+-- Detecting C compile features - done
+-- Detecting CXX compiler ABI info
+-- Detecting CXX compiler ABI info - done
+-- Check for working CXX compiler: /usr/bin/c++ - skipped
+-- Detecting CXX compile features
+-- Detecting CXX compile features - done
+-- No default build type specified: setting CMAKE_BUILD_TYPE=Release
+-- C++20 has been enabled by default
+-- Performing Test COMPILER_HAS_SSE3_FLAG
+-- Performing Test COMPILER_HAS_SSE3_FLAG - Success
+-- Performing Test COMPILER_HAS_AVX_FLAG
+-- Performing Test COMPILER_HAS_AVX_FLAG - Success
+-- Performing Test COMPILER_HAS_AVX2_FLAG
+-- Performing Test COMPILER_HAS_AVX2_FLAG - Success
+--
+-- PROJECT_NAME                = universal
+-- PROJECT_NAME_NOSPACES       = universal
+-- PROJECT_SOURCE_DIR          = /home/stillwater/dev/clones/universal
+-- PROJECT_VERSION             = 3.68.1.80df9073
+-- CMAKE_C_COMPILER            = /usr/bin/cc
[…]
