### 2023-02-06 #5072 @jamesquinlan
@yboettcher We could add more example templates to the paper, but probably readme best. Not sure if the length would be an issue if in the paper as I remember word ranges from https://joss.readthedocs.io/en/latest/submitting.html?highlight=length#what-should-my-paper-contain
