### 2023-02-06 #5072 @Ravenwater
@yboettcher I have incorporated your suggestions and added several presentations we have done through the years explaining the number systems contained. I also fixed all the bugs and misinformation you highlighted. This is all done in the v3.68 branch, please take a look and let us know if that satisfies your concerns.
