### 2023-02-07 stillwater-sc/universal#325 @mlxd
API documentation and use-case examples [JOSS #5072]
Hello, I am currently reviewing the library for as part of https://github.com/openjournals/joss-reviews/issues/5072

I see that several PDFs have been added to showcase the library (as in the [following link](https://github.com/stillwater-sc/universal/blob/8230030258521adbce0f72ecb52ae560d499a75f/docs/presentations/fptalks-custom-number-systems.pdf)). Though, I think the API-level documentation and overall guides are lacking for a developer to explore and use the library effectively.

Is there a plan to create a project-based site with a searchable API (Github pages, read-the-docs, etc), with guidelines and examples for usage? The linked PDF shows some detail, but this would be better served as part of a markdown (or equivalent) doc and linked to the front-level README.md for developers.

I think considering some like a doxygen-compatible CPP file headers and auto-generated API docs would also be beneficial here, for generating user-facing API docs throughout the repository.
