### 2023-02-07 #5072 @mlxd
Hi all, I have only a single item to add in addition to the comments from @yboettcher above:

I have added the above issue for item *Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?* https://github.com/stillwater-sc/universal/issues/325

I think some updates would be appreciated here, especially around visibility of function/modules, and guiding the developer accordingly with given examples.
