### 2023-02-09 #5072 @Ravenwater
@mlxd We Need to figure out how best to provide documentation. The types in _Universal_ are plug-in arithmetic types, so for regular use, there is no API as you would use the values in arithmetic expressions just like you would use a native float or int.

However, as @yboettcher also identified, documentation to bring folks up to speed with arithmetic types _Universal_ style would benefit from a description of how the collective was put together. That is more of a 'how-to-parameterize' arithmetic types how-to than API documentation, but would definitely be useful.

Then there is the whole verification environment and its design that does center around a common pattern to have the tests enumerate and drive raw bit patterns in the type so that the test framework can be generalized. If you look at the code allocation in _Universal_ you'll see that this is where the bulk of the code is, it is not in the arithmetic types.

Each type template has documentation of what the parameters mean. Do we want to bring this up in a documentation file?
