### 2023-02-09 1e4d48b
removing CodeShip CI as they stopped supporting open-source projects
