+++ CommandLineTools.md
@@ -1,17 +1,19 @@
 # Universal number system Command Line Tools
 
-The universal library contains a collection of command line tools that help investigate bit-level attributes of the number systems and the values they encode. These command line tools get installed with the `make install` build target.
+The universal library contains a collection of command line tools that help investigate 
+bit-level attributes of the number systems and the values they encode. These command line 
+tools get installed with the `make install` build target.
 
 The following segments presents short description of their use.
 
-## compieee
+## ieee
 
 Compare the three IEEE formats on a given real number value:
 
 ```text
-λ ./ieee.exe
+$ ieee
 Show the truncated value and (sign/scale/fraction) components of different floating point types.
-Usage: compieee floating_point_value
+Usage: ieee floating_point_value
 Example: ieee 0.03124999
 input value:                0.03124999
       float:              0.0312499907 (+,-6,11111111111111111111011)
@@ -19,135 +21,373 @@ input value:                0.03124999
 long double:  0.0312499899999999983247 (+,-6,111111111111111111101001011110100011111111111110001111111001111)
 ```
 
-## compf
+## quarter
 
-Show the sign/scale/fraction components of an IEEE float.
+Show the sign/scale/fraction components of an IEEE-754 quarter-precision floating-point.
 
 ```text
-λ ./float.exe
-compf : components of an IEEE single-precision float
-Show the sign/scale/fraction components of an IEEE float.
-Usage: float float_value
-Example: float 0.03124999
-float: 0.031249990686774254 (+,-6,11111111111111111111011)
+$ quarter
+quarter : components of an IEEE-754 quarter-precision float : 8 bits with 2 exponent bits
+Show the sign/scale/fraction components of a quarter-precision IEEE-754 floating-point.
+Usage: quarter value
+Example: quarter 0.03124999
+scientific   : 0.031
+triple form  : (+,  -5, 0b01.0'0000)
+binary form  : 0b0.00.0'0001
+color coded  : 00000001
+
+Number Traits of quarter-precision IEEE-754 floating-point
+std::numeric_limits< N2sw9universal6cfloatILj8ELj2EhLb1ELb0ELb0EEE >
+min exponent                                             -2
+max exponent                                              2
+radix                                                     2
+radix digits                                              6
+min                                                       1
+max                                                  3.9375
+lowest                                              -3.9375
+epsilon (1+1ULP-1)                                  0.03125
+round_error                                             0.5
+smallest value                                      0.03125
+infinity                                                inf
+quiet_NAN                                               nan
+signaling_NAN                                           nan
+
+smallest normal number
+0b0.01.00000 : 1
+smallest denormalized number
+0b0.00.00001 : 0.03125
 ```
 
-## compd
+## half
 
-Show the sign/scale/fraction components of an IEEE double.
+Show the sign/scale/fraction components of an IEEE-754 half-precision floating-point.
 
 ```text
-λ ./double.exe
-compd : components of an IEEE double-precision float
+$ half
+half : components of an IEEE-754 half-precision floating-point: 16 bits with 5 exponent bits
+Show the sign/scale/fraction components of a half-precision IEEE-754 floating-point.
+Usage: half value
+Example: half 0.03124999
+scientific   : 0.03125
+triple form  : (+,  -5, 0b01.00'0000'0000)
+binary form  : 0b0.0'1010.00'0000'0000
+color coded  : 0010100000000000
+
+Number Traits of half-precision IEEE-754 floating-point
+std::numeric_limits< N2sw9universal6cfloatILj16ELj5EtLb1ELb0ELb0EEE >
[…]
