+++ joss/paper.md
@@ -28,7 +28,7 @@ bibliography: references.bib
 
 # Summary
 
-*Universal Numbers Library*, or *Universal* for short, is a self-contained C++ header-only template library that contains implementations of many number representations and standard arithmetic on arbitrary configuration integer and real numbers [@omtzigt:2020].  In particular, the library includes integers, decimals, fixed-points, rationals, linear floats, tapered floats, logarithmic, SORNs, interval, level-index, and adaptive-precision binary and decimal integers and floats, each offering a verification suite.  
+*Universal Numbers Library*, or simply *Universal*, is a comprehensive, self-contained C++ header-only template library that provides implementations of various number representations and standard arithmetic operations on arbitrary configurations of integer and real numbers [@omtzigt:2020]. With its extensive collection of number systems, including integers, decimals, fixed-points, rationals, linear floats, tapered floats, logarithmic, SORNs, interval, level-index, and adaptive-precision binary and decimal integers and floats, Universal offers a robust verification suite for each system.
 
 The primary pattern using a posit number type as example, is:
 
@@ -51,17 +51,20 @@ int main() {
 }
 ```
 
-
 *Universal* delivers software and hardware co-design capabilities to develop low and mixed-precision algorithms for reducing energy consumption in signal processing, Industry 4.0, machine learning, robotics, and high-performance computing applications [@omtzigt:2022].  The package includes command-line tools for visualizing and interrogating numeric encodings, an interface for setting and querying bits, and educational examples showcasing performance gain and numerical accuracy with the different number systems.  In addition, a Docker container is available to experiment without cloning and building from the source code.
+
 
 ```
 $ docker pull stillwater/universal
 $ docker run -it --rm stillwater/universal bash
 ```
 
-*Universal* started in 2017 as a bit-level arithmetic reference implementation of the evolving unum Type III (posit and valid) standard.  However, the demands for supporting various number systems, such as adaptive-precision integers to solve large factorials, adaptive-precision floats to act as Oracles, or comparing linear and tapered floats provided the opportunity to create a complete platform for numerical analysis and computational mathematics.  As a result, several projects have leveraged *Universal*, including Matrix Template Library (MTL4), Geometry + Simulation Modules (G+SMO), Bembel, a fast IGA BEM solver, and the Odeint ODE solver.
+
+*Universal* was originally established in 2017 as a reference implementation of the evolving unum Type III (posit and valid) standard for bit-level arithmetic. However, as the demand for supporting a diverse range of number systems grew, *Universal* evolved into a complete platform for numerical analysis and computational mathematics, capable of solving problems such as large factorials using adaptive-precision integers and serving as Oracles using adaptive-precision floats. Many projects have leveraged *Universal*, including the Matrix Template Library (MTL4), Geometry + Simulation Modules (G+SMO), Bembel (a fast IGA BEM solver), and the Odeint ODE solver.
+
 
 The default build configuration will produce the command line tools, a playground, and educational and application examples.  It is also possible to construct the full regression suite across all the number systems.  For instance, the shortened output for the commands `single` and `single 1.23456789` are below.
+
 
 ```bash
 $ single
@@ -94,13 +97,14 @@ color coded  : 0b0.0111'1111.001'1110'0000'0110'0101'0010
 
 # Statement of need
 
[…]
