### 2023-02-09 bc0dbae
enhancing the documentation describing the parameterization of the arithmetic types
