### 2023-02-09 #5072 @Ravenwater
@mlxd I just checked in a new markup file called `docs/number-system-type-parameterization.md` that describes in a one pager how the core types are parameterized. This is in the v3.68 branch. Love to get your quick feedback if that satisfies your request.

In each number system type's regression suite there is always an `api/api.cpp` that chronicles all the invokation and use cases to provide an executable example of how to use the type. In general, the `api` section of the regression tests has code examples how to use the different library components, such as manipulators, attributes, number traits, exceptions, and special cases. Furthermore, iIn the `education` build target, there are individual test programs that demonstrate how to use the different types. The documentation might need to high light that these resources are available.
