+++ .readthedocs.yaml
@@ -1,13 +1,0 @@
-# File: .readthedocs.yaml
-
-version: 2
-
-sphinx:
-  configuration: docs/conf.py
-
-formats: all
-
-python:
-  version: 3.8
-  install:
-    - requirements: docs/requirements.txt
