### 2023-02-10 #5072 @Ravenwater
@mlxd @yboettcher @mmoelle1 I have updated the README.md to add extra sections that explain how to use and how to extend the library. This is in the v3.68 branch.

Love to hear your feedback if these additions alleviate your concerns.
