### 2023-02-11 #5072 @jamesquinlan
@danielskatz quick question, does setting to v3.68 read out of the v3.68 branch in the repo?  The article proof doesn't contain the revisions, e.g., the missing reference.

Thanks again for your help.
