### 2023-02-11 #5072 @danielskatz
No, this just sets the version tag. Changing the branch is a different command
