### 2023-02-11 stillwater-sc/universal#325 @Ravenwater
I have added a markdown one pager in docs and also added a segment to the
readme that explains the broad structure of the parameterization. Since
these are plug-in replacement types for native arithmetic types, the
constructor is the only API that is explicit, the main use cases are
implicit in the form of arithmetic and conversion expressions.
