### 2023-02-12 92d1da7
bug fix, incorrect computation bc transpose()
