### 2023-02-12 38348c6
bug fix in condition estimate (#328)
