### 2023-02-19 57884af
fix conjugate computation for lns and cfloat
