### 2023-02-19 #5072 @yboettcher
@jamesquinlan I like that paragraph much better now! The only phrasing I'm a bit confused by now is the "high energy consumption requirement" (l. 40). While I think it's totally clear what you want to say here, I have never seen "energy consumption requirement" in a single phrase. Only "energy requirement" or "energy consumption", so I feel like using both at the same time is a bit off.
