### 2023-02-22 #5072 @mlxd
> @mlxd @yboettcher @mmoelle1 I have updated the README.md to add extra sections that explain how to use and how to extend the library. This is in the v3.68 branch.
>
> Love to hear your feedback if these additions alleviate your concerns.

Hi @Ravenwater I think this helps address some concerns, though I think there would be benefit from tidying up stray XLSX and equivalent files in `docs`.
```
Documentation is provided via `Docs`, including a tutorial...
```
I think I'd also potentially suggest adding a directory level README.md to `docs` indicating the contents, and allowing a reader of the paper to easily find the intended tutorial as listed. Lastly, I'd match the case of `Docs` to that of the directory `docs` for consistency.

With these, I'm happy to give the all clear.
