### 2023-02-23 3e95712
restructuring the docs directory and adding more descriptive links
