+++ README.md
@@ -15,17 +15,20 @@
 -------------------------------
 
 
-The goal of the Universal Numbers Library is to offer applications alternatives to native integer and floating-point for mixed-precision algorithm development and optimization. Tailoring the arithmetic types to the application's precision and dynamic range requirements enables a new level of performance and energy efficiency, valuable for embedded applications that require high computational density and efficiency.
+The goal of the Universal Numbers Library is to offer alternatives to native integer and floating-point for mixed-precision algorithm development and optimization. Tailoring the arithmetic types to the application's precision and dynamic range enables a new level of application performance and energy efficiency, particularly valuable for embedded applications that need autonomy through intelligent behavior.
 
-Deep Learning algorithms in particular provide a core application vertical where alternative formats and precisions, such as half-precision floating-point and bfloat16s yield speed-ups of two to three orders of magnitude, making rapid innovation in AI possible.
+Deep Learning algorithms in particular provide a core application vertical where alternative formats and precisions, such as half-precision floating-point and bfloat16, yield speed-ups of two to three orders of magnitude, making rapid innovation in AI possible.
 
-The Universal Library is a ready-to-use header-only library that provides a plug-in replacement for native types and a low-friction environment to explore alternatives to IEEE-754 floating-point in algorithms.
+The Universal Library is a ready-to-use header-only library that provides a plug-in replacement for native types and offers a low-friction environment to explore alternatives to IEEE-754 floating-point in AI, DSP, HPC, and HFT algorithms.
 
 The basic use pattern is as simple as:
 
 ```code
+// bring in the parameterized type of interest, in this case
+// a fixed-sized, arbitrary configuration classic floating-point
 #include <universal/number/cfloat/cfloat.hpp>
 
+// define your computational kernel parameterized by arithmitic type
 template<typename Real>
 Real MyKernel(const Real& a, const Real& b) {
     return a * b;  // replace this with your kernel computation
@@ -34,10 +37,12 @@ Real MyKernel(const Real& a, const Real& b) {
 constexpr double pi = 3.14159265358979323846;
 
 int main() {
+    // if desired, create an application type alias to avoid errors
     using Real = sw::universal::half; // half-precision IEEE-754 floating-point  
 
     Real a = sqrt(2);
     Real b = pi;
+    // finally, call your kernel with your desired arithmetic type
     std::cout << "Result: " << MyKernel(a, b) << std::endl;  
 }
 ```
@@ -54,7 +59,7 @@ The library contains fast implementations of special IEEE-754 formats that do no
 
 ## Citation
 
-Please cite [our work](https://arxiv.org/abs/2012.11011) if you use _Universal_.
+Please cite [our work](https://arxiv.org/abs/2012.11011) if you use *Universal*.
 
 ```bib
 @inproceedings{Omtzigt:2022,
@@ -76,7 +81,7 @@ Please cite [our work](https://arxiv.org/abs/2012.11011) if you use _Universal_.
 
 ## Talks and Presentations
 
-The following presentations describe _Universal_ and the number systems it contained as of the time of publication.
+The following presentations describe *Universal* and the number systems it contained as of the time of publication.
 
 Slides of a presentation at [FPTalks'21](https://fpbench.org/talks/fptalks21.html)
 
@@ -101,7 +106,7 @@ CMakeFiles           applications  cmake_install.cmake    tests       universal-
 CTestTestfile.cmake  c_api         education              tools       universal-targets.cmake
 ```
 
-[Here](docs/CommandLineTools.md) is a quick reference of what the command line tools have to offer.
+[Here](docs/command-line-tools.md) is a quick reference of what the command line tools have to offer.
 
 ## How to build
 
[…]
