+++ docs/README.md
@@ -1,0 +1,38 @@
+# Directory structure of `docs`
+
+Tutorials: 
+1. [command line tools description](command-line-tools.md)
+2. [number type parameterization](number-system-type-parameterization.md)
+
+## deep_learning
+
+Contains a spreadsheet that compares different 8-bit arithmetic types.
+
+## img
+
+Contains images used in the markup and tutorials
+
+## presentations
+
+Contains presentation slides:
+1. [CoNGA22](presentations/conga22-universal-arithmetic-library.pdf)
+2. [FPTalks](fptalks-custom-number-systems.pdf)
+
+## ranges
+
+Contains the dynamic ranges of a select set of classic floats, 
+fixpnts, posits, and lns types.
+
+## references
+
+Contains tidbits of information to guide computational mathemetical
+problems in function evaluation.
+
+## regressions
+
+Contain historical regression results
+
+## tables
+
+Contains ready to use value tables of small classic floats,
+fixpnts, posits, and lns types.
