### 2023-02-23 #5072 @Ravenwater
@mlxd Thank you for that feedback. I have cleaned up the `docs` directory, added the docs/README.md that explains the content, and added a description of al the new  `docs` directory goodies in the repo README in the section *How to develop and extend Universal".

It is this commit: https://github.com/stillwater-sc/universal/commit/3e957128214920af1fe4a2bf79bb9577f034de31
