### 2023-02-23 #5072 @mlxd
> @mlxd Thank you for that feedback. I have cleaned up the `docs` directory, added the docs/README.md that explains the content, and added a description of al the new `docs` directory goodies in the repo README in the section *How to develop and extend Universal".
>
> It is this commit: [stillwater-sc/universal@3e95712](https://github.com/stillwater-sc/universal/commit/3e957128214920af1fe4a2bf79bb9577f034de31)

Thanks @Ravenwater I'll resolve the issue https://github.com/stillwater-sc/universal/issues/325 and consider my review checklist complete. Nothing more from my side, and great job on the work in this library.
