### 2023-02-26 ed6cc4e
adding posit visualization back into the docs
