### 2023-03-02 26bf524
add integrations (commented out for some types) into cmath functions
