+++ include/universal/number/efloat/efloat.hpp
@@ -35,4 +35,7 @@
 /// math functions
 //#include <universal/number/efloat/math_functions.hpp>
 
+// no enable_if_efloat
+// #include <universal/number/efloat/cmath_integration.hpp>
+
 #endif
