+++ include/universal/number/rational/cmath_integration.hpp
@@ -14,6 +14,7 @@ namespace std {
     make_std_fun(exp)
     make_std_fun(log)
     make_std_fun(atan)
+    make_std_fun(tan)
 
     template <typename T>
     sw::universal::enable_if_rational<T, std::string> to_string(const T &t) {
