### 2023-03-04 #5072 @yboettcher
For me, the only thing left to do is (successfully) testing out the different number formats offered by Universal. I wanted to do this like I used the posits back when I wrote my thesis, but most number formats other than posit and fixpnt resulted in compile-time errors (which I then turned into a fix [here](https://github.com/stillwater-sc/universal/pull/329) and a proposed solution [here](https://github.com/stillwater-sc/universal/pull/330)). This might however be because I was using Eigen, while the software paper and universal docs only mention MTL4. So trying out the number formats with MTL4 would be my next (and then also last) step.
A last thing I would mention, although that's totally not a blocker: Universal offers SORNs, which is also mentioned by the software paper, but I had a hard time figuring out, what SORNs would be (other than "sets of real numbers"). I eventually found this https://link.springer.com/chapter/10.1007/978-3-031-09779-9_1, but I thought maybe some reference (maybe not even in the paper and just in the Readme) would be great. I would also be fine though if it is deemed out-of-scope and a user is expected to know what they want.
