### 2023-03-05 #5072 @yboettcher
I think I would like some guidance here. I uploaded my test cases to https://github.com/yboettcher/universal_tests with most of them being disabled in the CMakeLists.txt. The only one, where I can compile most of the number formats and actually test them is a simple test for +-*/. Everything involving Eigen/MTL does not compile with most number formats for me. For SORNs (which are mentioned as supported in the software paper) I cannot compile this simple test, as the compiler complains about missing implementations for +-*/.
Now, I admit that throwing these types at Eigen/MTL and expecting them to work fine might be a bit too much (after all, especially with Eigen, no one ever claimed there was any compatibility with it. It worked very good for posits though).
Now if it were just for the posits: I have used them extensively with Eigen and Bembel and would happily check off any remaining items on the checklist. However, I was trying to gauge all the other formats that universal offers and was surprised they did not work so seamlessly. I think my overarching question here is: Should I check off the functionality (and performance), because universal's main offering, the posits,  have worked so well for me, or should I insist on all the other types working too, because the software paper indeed states that universal offers support for them? (once again, not working with Eigen is fine by me, because it never claimed to do so, and someone using universal would just have to do linear algebra manually, but the thing I am most surprised about is most likely the SORNs not working for me at all, although they were explicitely mentioned in the software paper.)
As a heads up, the other important number formats (posits, cfloat, logarithmic, and some more) pass this +-*/ test and are fine on my side. There were some other failing tests, but I am not too worried about those.
