### 2023-03-05 #5072 @theo-lemurian
@danielskatz @yboettcher I am investigating the Eigen breakage and the PRs Yannick provided. Folks that use Eigen. like GISMO, have typically provided Eigen-specific extensions for their integrations, but I will try to see if we can close this feature if it is isolated to Eigen asking for a specific constructor.

SORNs are provided by a research team from Moritz Bartel from University of Bremen: https://link.springer.com/chapter/10.1007/978-3-031-09779-9_1  and as their paper indicates, SORNs are not to be used as plug-in replacement arithmetic as they exploit very specific behaviors from the encoding, these SORNs are not a general arithmetic.
