### 2023-03-05 #5072 @theo-lemurian
@yboettcher  the compilation error you encounter is correct behavior. The standard library provides functions for the native types, so `std::sqrt(someNativeType)` will find the sqrt() function. However, if you want to call a specialized function for an extension type, you will need to call the `sqrt()` function defined for that specialization. For *Universal* these functions reside in the namespace `sw::universal`. Thus the pattern that is correct is to define your uses like this:

```
namespace yourNamespace {
  template<typename Scalar>
  Scalar myKernel() {
    using std::sqrt;
    Scalar a = 1507.2; // arbitrary number
    Scalar b = sqrt(a); // arbitrary function from
    return a * a;
  }
}
```

The `using std::sqrt` is required to match the `std` namespace functions in case the Scalar is a native type. When the Scalar type is a *Universal* type, the compiler will be able to infer the namespace it needs to match the argument.

So, your test environment is not using the proper C++ patterns to use *Universal*.
