### 2023-03-08 #5072 @jamesquinlan
@yboettcher @danielskatz Any thoughts on the latest remarks?  Do you agree with @Ravenwater analysis?  Thanks
