### 2023-03-14 4441251
fix conjugate computation for lns and cfloat (#329)
fixing conj() for lns and cfloat
