### 2023-03-14 43e9323
V3.68 (#331)
* bumping SEMVER to v3.68

* generalizing Dockerfile

* streamlining build, test, dev, and release containers

* adding deep learning floating-point formats

* removing the static_assert for long double being an extended precision floating-point

* removing CodeShip CI as they are no longer supporting open-source projects

* moving to checkout/actions@v3

* cody hygiene for edge case es == 0

* fixing up README using JOSS reviewer feedback

* adding published presentations talking about Universal and its number systems

* file name fix

* removing CodeShip CI as they stopped suppo
[…]
