### 2023-03-14 #5072 @Ravenwater
> # Post-Review Checklist for Editor and Authors
> ## Additional Author Tasks After Review is Complete
> * [x]  Double check authors and affiliations (including ORCIDs)
> * [x]  Make a release of the software with the latest changes from the review and post the version number here. *v3.68* This is the version that will be used in the JOSS paper.
> * [x]  Archive the release on Zenodo/figshare/etc and post the DOI here. 10.5281/zenodo.7735084
> * [x]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
>
