### 2023-03-14 f57675d
changing Docs to docs in JOSS paper
