### 2023-03-14 #5072 @danielskatz
Proofreading the paper, I see three issues:

I don't understand the grammar in line 33
`using a posit number type as example, is`

Similarly in line 54, something seems wrong in
`a Docker container is available to experiment without cloning and building from the source code.`

Finally, in line 142, please add a comma after `i.e.` so that it becomes `i.e.,`
