### 2023-03-14 7c4f697
final JOSS paper edits suggested by editor
