### 2023-03-14 #5072 @Ravenwater
@danielskatz

Line 33 changed to:
`Using a posit number as example, the basic pattern to use a custom *Universal* type is:`

Line 54 changed to:
`Finally, Docker containers are available to experiment with the library without cloning the repo and building the source code.`

Added comma after `i.e.` on line 142.
