### 2023-03-14 #5072 @danielskatz
In 33, should this be "as an example" or "for example"? I still don't understand the grammar.

The other 2 changes seems fine
