### 2023-03-15 #5072 @danielskatz
When it says to wait a few minutes after this finishes, please do wait at least 5 minutes before checking the DOI to avoid DNS caching issues.  I'll close the issue when this is process is complete
