### 2023-03-15 #5072 @Ravenwater
@danielskatz The JOSS process is absolutely brilliant, and the tooling around it is impressive. This is the first publishing framework that actually makes the paper better with very tight and continuous reviewer/editor feedback.
